Split canonical storage models
Move canonical ORM definitions into focused base, profile, content,
history, and reference model modules so each file stays under the
project line-count limit.
Keep `storage.models` as the Alembic-facing compatibility re-export.
