+++ episodic/canonical/storage/reference_repositories.py
@@ -20,11 +20,6 @@ from episodic.canonical.reference_protocols import (
     ReferenceDocumentRevisionRepository,
 )
 
-from .models import (
-    ReferenceBindingRecord,
-    ReferenceDocumentRecord,
-    ReferenceDocumentRevisionRecord,
-)
 from .reference_mappers import (
     _reference_binding_from_record,
     _reference_binding_to_record,
@@ -32,6 +27,11 @@ from .reference_mappers import (
     _reference_document_revision_from_record,
     _reference_document_revision_to_record,
     _reference_document_to_record,
+)
+from .reference_models import (
+    ReferenceBindingRecord,
+    ReferenceDocumentRecord,
+    ReferenceDocumentRevisionRecord,
 )
 from .repository_base import _RepositoryBase
